mainApp/requirements.txt
asgiref==3.2.7
Django==3.0.5
django-crispy-forms==1.9.1
lxml==4.5.0
mysqlclient==1.4.6
Pillow==7.1.2
python-docx==0.8.10
python-pptx==0.6.18
pytz==2019.3
sqlparse==0.3.1
XlsxWriter==1.2.8